changed file structure to make more sense
