--- Assets/FlashCards.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 58b4057e4792659d093f8b8b03fbf725
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
