--- Assets/FlashCards/ESCMenu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 395b9664014b641cf94d723c48e43ca8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
