--- Assets/FlashCards/FlashCardButtons.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 43fc2fe0b664415f184bc8cbca7982fd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
