--- Assets/FlashCards/FlashCardButtons/FailButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0330d4092bdf8c29e8c65dd54eab33b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
