--- Assets/FlashCards/FlashCardButtons/PassButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68f9c26f7e948a606a37352519810fc9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
