+++ Assets/Src/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed55f09e393d2b344b300d1a166c57e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
