+++ Assets/Src/Scripts/FlashCards/ESCMenu/EscapeMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3091a3cbb139b710b94cc2cd91aa6e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
