+++ Assets/Src/Scripts/FlashCards/ESCMenu/ResumeFromESCPMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d02fc2646ec573e61b5339ffaf0135b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
