+++ Assets/Src/Scripts/FlashCards/EnemyManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfc710b4e4ced79eca2880b5f7a01937
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
