+++ Assets/Src/Scripts/FlashCards/FlashCardButtons/CardManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3c43bad76546ec6c9a4e308001fd6f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
