+++ Assets/Src/Scripts/FlashCards/FlashCardButtons/ContinueButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90159064332cc3d6dbcb5962cc934ec9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
