+++ Assets/Src/Scripts/FlashCards/Kanji.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2259e2b20d27dad1cb3a2a54fe7018a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
