+++ Library/Bee/artifacts/2400b0aE.dag/Assembly-CSharp.rsp
@@ -348,15 +348,15 @@
 -r:"Library/Bee/artifacts/2400b0aE.dag/Unity.VisualStudio.Editor.ref.dll"
 -r:"Library/Bee/artifacts/2400b0aE.dag/UnityEditor.UI.ref.dll"
 -r:"Library/Bee/artifacts/2400b0aE.dag/UnityEngine.UI.ref.dll"
-"Assets/FlashCards/CardManager.cs"
-"Assets/FlashCards/ESCMenu/EscapeMenu.cs"
-"Assets/FlashCards/ESCMenu/ResumeFromESCPMenu.cs"
-"Assets/FlashCards/EnemyManager.cs"
-"Assets/FlashCards/FlashCardButtons/ContinueButton.cs"
-"Assets/FlashCards/FlashCardButtons/FailButton.cs"
-"Assets/FlashCards/FlashCardButtons/PassButton.cs"
-"Assets/FlashCards/Kanji.cs"
 "Assets/SU3DJPFont/Demo/SampleScene.cs"
+"Assets/Src/FlashCards/ESCMenu/EscapeMenu.cs"
+"Assets/Src/FlashCards/ESCMenu/ResumeFromESCPMenu.cs"
+"Assets/Src/FlashCards/EnemyManager.cs"
+"Assets/Src/FlashCards/FlashCardButtons/CardManager.cs"
+"Assets/Src/FlashCards/FlashCardButtons/ContinueButton.cs"
+"Assets/Src/FlashCards/FlashCardButtons/FailButton.cs"
+"Assets/Src/FlashCards/FlashCardButtons/PassButton.cs"
+"Assets/Src/FlashCards/Kanji.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark01.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark01_UGUI.cs"
 "Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark02.cs"
